use warcio recompress as a syntax check

+++ tests/warc/test.sh
@@ -10,12 +10,19 @@ echo
 echo test-warc
 echo
 
+# pre-cleanup
+rm -f CC-TEST-01-FOO-00001-*.warc
+rm -f CC-TEST-01-FOO-00002-*.warc
+rm -f CC-TEST-01-00001-*.warc.g
+
 $COVERAGE ./test-warc.py
 
 # there are 3 output files to diff... move them to canonical names
-mv -u CC-TEST-01-FOO-00001-*.warc CC-TEST-01-FOO-00001-hostname.warc
-mv -u CC-TEST-01-FOO-00002-*.warc CC-TEST-01-FOO-00002-hostname.warc
-mv -u CC-TEST-01-00001-*.warc.gz CC-TEST-01-00001-hostname.warc.gz
+mv CC-TEST-01-FOO-00001-*.warc CC-TEST-01-FOO-00001-hostname.warc
+mv CC-TEST-01-FOO-00002-*.warc CC-TEST-01-FOO-00002-hostname.warc
+mv CC-TEST-01-00001-*.warc.gz CC-TEST-01-00001-hostname.warc.gz
+
+# diffs
 
 $COVERAGE ./warc-diff.py CC-TEST-01-FOO-00001-hostname.warc.in CC-TEST-01-FOO-00001-hostname.warc
 $COVERAGE ./warc-diff.py CC-TEST-01-FOO-00002-hostname.warc.in CC-TEST-01-FOO-00002-hostname.warc
@@ -25,6 +32,15 @@ gunzip -S .gz.in --to-stdout CC-TEST-01-00001-hostname.warc.gz.in > CC-TEST-01-0
 rm -f CC-TEST-01-00001-hostname.warc
 gunzip CC-TEST-01-00001-hostname.warc.gz
 $COVERAGE ./warc-diff.py CC-TEST-01-00001-hostname.warc.in CC-TEST-01-00001-hostname.warc
+
+# recompress to see if the syntax is happy
+rm -f foo.warc.gz
+warcio recompress CC-TEST-01-00001-hostname.warc foo.warc.gz
+if [ `warcio index CC-TEST-01-00001-hostname.warc | wc -l` != `warcio index foo.warc.gz | wc -l` ]; then
+    echo "warc index size changed on recompress"
+    exit 1
+fi
+rm -f foo.warc.gz
 
 # cleanup
 rm CC-TEST-01-FOO-00001-hostname.warc
